CPE_Development
Update to the project, added a function to collect the CPE dataset using the API and multithreading to collect 1.2 million records in about 3 minutes. The results are then inserted into a sqlite3 database.

Added a dummy cpe.ttl file to try and update the ontology, added functions (try_call) and (check_status) to maintain code readability.

Imported concurrent.futures to run multiple api_calls at the same time. In total, about 250+ lines of code were added. Pending items are to add the CPE to the ontology.